// src/types/enums.ts
export enum Environment {
    DEVELOPMENT = "development",
    PRODUCTION = "production",
}

export enum ErrorLabel {
  // -------------------- AUTH --------------------
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  INVALID_TOKEN = "INVALID_TOKEN",
  TOKEN_EXPIRED = "TOKEN_EXPIRED",
  UNAUTHORIZED = "UNAUTHORIZED",
  FORBIDDEN = "FORBIDDEN",

  // ----------------- VALIDATION -----------------
  BAD_REQUEST = "BAD_REQUEST",
  VALIDATION_ERROR = "VALIDATION_ERROR",
  INVALID_INPUT = "INVALID_INPUT",
  INVALID_EMAIL = "INVALID_EMAIL",
  WEAK_PASSWORD = "WEAK_PASSWORD",

  // --------------------- USER -------------------
  USER_NOT_FOUND = "USER_NOT_FOUND",
  USER_EXISTS = "USER_EXISTS",
  ACCOUNT_INACTIVE = "ACCOUNT_INACTIVE",
  ACTION_NOT_ALLOWED = "ACTION_NOT_ALLOWED",

  // ------------------- DATABASE -----------------
  NOT_FOUND = "NOT_FOUND",
  DUPLICATE_ENTRY = "DUPLICATE_ENTRY",
  DB_CONNECTION_ERROR = "DB_CONNECTION_ERROR",
  DB_QUERY_ERROR = "DB_QUERY_ERROR",
  DB_TIMEOUT = "DB_TIMEOUT",

  // --------------------- SYSTEM -----------------
  INTERNAL_ERROR = "INTERNAL_ERROR",
  UNKNOWN_ERROR = "UNKNOWN_ERROR",
  EXTERNAL_SERVICE_ERROR = "EXTERNAL_SERVICE_ERROR",
  CACHE_ERROR = "CACHE_ERROR",
  FILE_SYSTEM_ERROR = "FILE_SYSTEM_ERROR",

  // ------------------ FILE UPLOAD ---------------
  FILE_TOO_LARGE = "FILE_TOO_LARGE",
  INVALID_FILE_TYPE = "INVALID_FILE_TYPE",
  UPLOAD_FAILED = "UPLOAD_FAILED",

  // --------------- COMMUNICATION ----------------
  EMAIL_ERROR = "EMAIL_ERROR",
  SMS_ERROR = "SMS_ERROR",
  WEBHOOK_ERROR = "WEBHOOK_ERROR",
}

export enum HttpStatus {
  // ----- 2xx SUCCESS -----
  OK = 200,
  CREATED = 201,
  ACCEPTED = 202,
  NO_CONTENT = 204,

  // ----- 3xx REDIRECTION -----
  MOVED_PERMANENTLY = 301,
  FOUND = 302,
  NOT_MODIFIED = 304,

  // ----- 4xx CLIENT ERRORS -----
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  METHOD_NOT_ALLOWED = 405,
  CONFLICT = 409,
  PAYLOAD_TOO_LARGE = 413,
  UNSUPPORTED_MEDIA_TYPE = 415,
  TOO_MANY_REQUESTS = 429,

  // ----- 5xx SERVER ERRORS -----
  INTERNAL_SERVER_ERROR = 500,
  NOT_IMPLEMENTED = 501,
  BAD_GATEWAY = 502,
  SERVICE_UNAVAILABLE = 503,
  GATEWAY_TIMEOUT = 504,
}